backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
jieba==0.42.1
jsonlines==4.0.0
ollama==0.6.1
pydantic>=2.12.4
PyYAML==6.0.3
rank_bm25==0.2.2
tqdm==4.67.1
faiss-cpu==1.13.0
nltk
pysbd
rouge
rouge_chinese
openai
langchain
psycopg2-binary
